code refactor #53
