+++ app/src/main/java/com/example/moodtrackr/MainActivity.kt
@@ -11,6 +11,8 @@ import androidx.activity.compose.setContent
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
+import com.example.moodtrackr.screens.EditProfileScreen
+import com.example.moodtrackr.screens.HomeScreen
 
 // MainActivity.kt
 class MainActivity : ComponentActivity() {
