+++ app/src/main/java/com/example/moodtrackr/models/Profile.kt
@@ -1,0 +1,5 @@
+package com.example.moodtrackr.models
+
+class Profile(var name: String, var surname: String,
+              var sex: String, var birthday: String) {
+}
