+++ app/src/main/java/com/example/moodtrackr/models/SharedPreferencesConstants.kt
@@ -1,4 +1,4 @@
-package com.example.moodtrackr
+package com.example.moodtrackr.models
 
 object SharedPreferencesKeys {
     const val PROFILE = "Profile"
