+++ app/src/main/java/com/example/moodtrackr/repositories/IProfilePreferenceManager.kt
@@ -1,0 +1,8 @@
+package com.example.moodtrackr.repositories
+
+import com.example.moodtrackr.models.Profile
+
+interface IProfilePreferenceManager {
+    fun load() : Profile
+    fun save(profile : Profile)
+}
